backend/requirements.txt
autopep8==1.5.4
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
fake-useragent==0.1.11
firebase-admin==4.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.4
google-cloud-firestore==2.0.2
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
httmock==1.4.0
httplib2==0.18.1
idna==2.10
msgpack==1.0.0
proto-plus==1.11.0
protobuf==3.14.0
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
rope==0.18.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
